+++ longevity/src/main/scala/longevity/config/MongoConfig.scala
@@ -1,4 +1,4 @@
-package longevity.context
+package longevity.config
 
 /** the mongo configuration
  *