+++ longevity/src/main/scala/longevity/config/package.scala
@@ -1,0 +1,4 @@
+package longevity
+
+/** contains the [[LongevityConfig]] plus supporting types and classes */
+package object config