+++ longevity/src/test/scala/longevity/integration/poly/CassandraPolyReposSpec.scala
@@ -2,7 +2,7 @@ package longevity.integration.poly
 
 import longevity.ConfigMatrixKey
 import longevity.TestLongevityConfigs
-import longevity.context.Cassandra
+import longevity.config.Cassandra
 import longevity.context.LongevityContext
 import longevity.integration.subdomain.derived