+++ longevity/src/test/scala/longevity/integration/writeConflict/InMemOptLockSpec.scala
@@ -3,7 +3,7 @@ package longevity.integration.writeConflict
 import longevity.ConfigMatrixKey
 import longevity.TestLongevityConfigs
 import longevity.context.LongevityContext
-import longevity.context.InMem
+import longevity.config.InMem
 import longevity.integration.subdomain.basics
 
 /** optimistic locking tests for inmem back end */